fixed percentages with float numbers

+++ spec/models/batch_spec.rb
@@ -23,7 +23,7 @@ RSpec.describe Batch, type: :model do
     let(:evaluation) { create :evaluation, color: "green", student: student4 }
 
       it "returns percentage of students with the color red, yellow or blue" do
-        expect(batch.return_percentage("red")).to be_a(integer)
+        expect(batch.return_percentage("red")).to be_a(Float)
       end
   end
